modify authrity

+++ core/src/main/java/com/ifhz/core/po/Role.java
@@ -39,7 +39,7 @@ public class Role {
     /**
      * 层级
      */
-    private Integer levels;
+    private Long levels;
 
     /**
      * 创建时间
@@ -79,11 +79,11 @@ public class Role {
         this.fullPath = fullPath;
     }
 
-    public Integer getLevels() {
+    public Long getLevels() {
         return levels;
     }
 
-    public void setLevels(Integer levels) {
+    public void setLevels(Long levels) {
         this.levels = levels;
     }
